requirements.txt
asgiref==3.7.2
Django==4.2.2
django-tinymce==3.6.1
python-decouple==3.8
sqlparse==0.4.4
